requirements.txt
arabic-reshaper==3.0.0
asgiref==3.8.1
asn1crypto==1.5.1
Brotli==1.1.0
certifi==2024.12.14
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
cryptography==44.0.0
cssselect2==0.7.0
Django==5.1.2
django-jazzmin==3.0.1
fonttools==4.54.1
html5lib==1.1
idna==3.10
lxml==5.3.0
mysqlclient==2.2.5
oscrypto==1.3.0
pillow==10.4.0
pycparser==2.22
pydyf==0.11.0
pyHanko==0.25.3
pyhanko-certvalidator==0.26.5
pypdf==5.1.0
pyphen==0.15.0
python-barcode==0.15.1
python-bidi==0.6.3
python-decouple==3.8
PyYAML==6.0.2
qrcode==8.0
reportlab==4.2.5
requests==2.32.3
six==1.17.0
sqlparse==0.5.1
svglib==1.5.1
tinycss2==1.4.0
tinyhtml5==2.0.0
typing_extensions==4.12.2
tzdata==2024.2
tzlocal==5.2
uritools==4.0.3
urllib3==2.2.3
weasyprint==63.0
webencodings==0.5.1
xhtml2pdf==0.2.16
zopfli==0.2.3.post1
